fix(FR-2520): use currentProject instead of model-store project in Model Card deploy (#6593)
Resolves #6590 (FR-2520)

## Summary

Two related defects in `ModelCardDeployModal` that prevented regular (non-admin) users from deploying from a Model Card:

1. **Wrong project context for deploy** — `useModelStoreProject()` (the MODEL_STORE marketplace project) was being passed as `projectId` to `deployModelCardV2`. The spec requires the user's current working project. Fixed by switching to `useCurrentProjectValue()`.
2. **Empty resource group selector** — the modal queried the admin-only top-level `scaling_groups` GraphQL field with no project filter, so regular users saw an empty list and could not deploy. Fixed by reusing the same project-scoped REST endpoint as `BAIProjectResourceGroupSelect` (`/scaling-groups?group=<projectName>`), which already powers session creation for regular users.

## Changes

- **New** `packages/backend.ai-ui/src/hooks/useProjectResourceGroups.ts` — extracted shared hook that fetches project-scoped resource groups via the REST endpoint `/scaling-groups?group=<projectName>`, filters out SFTP-only scaling groups, and exposes the result. Single source of truth so the modal and the select component cannot disagree about what's available.
- `packages/backend.ai-ui/src/components/BAIProjectResourceGroupSelect.tsx` — refactored to consume `useProjectResourceGroups`. No behavior change for existing call sites (same query key, same filtering, same React Query cache key).
- `packages/backend.ai-ui/src/hooks/index.ts` — export the new hook and its `ScalingGroupItem` type.
- `react/src/components/ModelCardDeployModal.tsx`:
    - Use `useCurrentProjectValue()` for `projectId` in the `deployModelCardV2` mutation (fix #1).
    - Use `useProjectResourceGroups(projectName)` to fetch the resource group list — same React Query key as the select component, so only one network call is made (fix #2).
    - Replace the inline `BAISelect` for the resource group with `BAIProjectResourceGroupSelect`, which already handles SFTP filtering, auto-default selection, and search.
    - Drop `scaling_groups` from `ModelCardDeployModalQuery` (no longer needed).
    - Auto-deploy scenario detection (`availablePresets.length === 1 && resourceGroups.length === 1`) is preserved by reading `resourceGroups` from the shared hook.

## Verification

- `bash scripts/verify.sh` → `=== ALL PASS ===` (Relay, Lint, Format, TypeScript)

## Spec reference

`.specs/model-card-detail-deploy-ux/spec.md:184` — projectId defaults to the user's current working project; no project-selection UI is included in the deploy modal.

## Stack context

This fix is stacked on top of `04-09-test_fr-2503_add_e2e_tests_for_model_card_drawer_and_deploy_ux`, which transitively includes PR #6517 (`feat/FR-2489-deploy-ux`) where both bugs were introduced.
